+++ udata_piwik/models.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from datetime import datetime
+
+from udata.models import db
+
+
+__all__ = ('PiwikTracking', )
+
+
+class PiwikTracking(db.Document):
+    url = db.StringField(required=True)
+    date = db.DateTimeField(required=True, default=datetime.now)
+    kwargs = db.DictField()
+
+    meta = {
+        'indexes': ['date'],
+        'ordering': ['date'],
+    }